feat: create keycloak callback page
