+++ packages/app/package.json
@@ -26,7 +26,7 @@
     "@nl-portal/nl-portal-user-interface": "workspace:*",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.25.1"
+    "react-router-dom": "^6.27.0"
   },
   "devDependencies": {
     "@types/jest": "^29.5.12",
