+++ packages/authentication/package.json
@@ -43,6 +43,7 @@
   "peerDependencies": {
     "@react-keycloak/web": "^3.4.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.27.0"
   }
 }
