+++ packages/authentication/src/index.ts
@@ -1,3 +1,4 @@
 export { default as KeycloakWrapper } from "./components/KeycloakWrapper";
 export { default as KeycloakContext } from "./contexts/KeycloakContext";
+export { default as KeycloakCallbackPage } from "./pages/KeycloakCallbackPage";
 export * from "./utils/format-url-trailing-slash";
