+++ packages/user-interface/package.json
@@ -101,6 +101,6 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-intl": "^6.6.6",
-    "react-router-dom": "^6.23.1"
+    "react-router-dom": "^6.27.0"
   }
 }
